requirements.txt
fastapi
uvicorn
jinja2
python-multipart
apscheduler
boto3
psycopg2-binary
python-dotenv
httpx
